==> 0_Overview.Rmd <==
---
title: "COVID-19 Pandemic and Olympics Marathon Performance - A Case Study Using Rio 2016 and Tokyo 2020 Data -"
author: "BST260 Group 11: Mariko Ando, Yi-Ting Tsai, Satoko Ugai, Marie Wu, Yusuke Yoshikawa, Yuki Zhong"
output: html_document
---

## 0. Overview (BST260 Group 11 Members)
### OVERVIEW AND MOTIVATION

COVID-19 has impacted the world over, without sparing a single individual or country. It has been pointed out in the literature that the COVID-19 pandemic impacted not only those infected by COVID-19, but also the healthy population. Previous questionnaire surveys have reported that obesity, depression, and anxiety increased in healthy people during COVID-19 pandemic. Regional or nation-wide lockdowns and modification of work mode have resulted in decreased levels of physical activity in the healthy population. Since physical exercise is significantly associated with mental health and obesity, lack of exercise can lead to higher chances of developing obesity and depression. 

   Meanwhile, on the national or international level, concerns of viral transmission have caused the cancellation or postponement of various events, such as the 2020 Tokyo Summer Olympics. Originally planned to be held in the summer of 2020, these Olympic games were postponed indefinitely in March 2020. Athletes faced interrupted training routines, as many countries implemented restrictive lockdowns and shut down gyms and athlete training centers. Some athletes reportedly contracted the virus or were even hospitalized. It arouses our interest to explore the association between the pandemic and athlete performance.
   
### RELATED WORKS
The impact of COVID-19 on healthy population: 

  1. [Changes in Weight and Nutritional Habits in Adults with Obesity during the "Lockdown" Period Caused by the COVID-19 Virus Emergency](http://m.shengsci.com/wenxian/631617.html) 
  
  2. [Prevalence and Influencing Factors of Anxiety and Depression Symptoms in the First-Line Medical Staff Fighting Against COVID-19 in Gansu](https://www.frontiersin.org/articles/10.3389/fpsyt.2020.00386/full) 
  
  3. [Association between physical exercise and mental health in 1·2 million individuals in the USA between 2011 and 2015: a cross-sectional study](https://pubmed.ncbi.nlm.nih.gov/30099000/) 
  
About the Tokyo 2020 Olympics: 

  1. [Blame the pandemic. Or the morning finals. But these Olympic swim times are slow.](https://www.washingtonpost.com/sports/olympics/2021/07/26/olympics-swimming-slow-times-pandemic-tokyo/)
  
  2. [A look at all of the world records that were broken at the Tokyo Olympics.](https://www.nytimes.com/2021/08/08/sports/olympics/world-records-tokyo-olympics.html)
  

### INITIAL QUESTIONS

The primary scientific goal of our project is to see whether there was any difference in records for marathon games between pre- and post-pandemic periods. Our secondary scientific goal is to investigate whether countries with different COVID-19 severity showed different trends for a change in records between pre- and post-pandemic periods. Over the course of the project, we found we would also like to investigate whether countries with different GDPs showed difference in the marathon records, since the economic status during the pandemic could affect the attendance of the Olympics. By doing these analyses, we can evaluate the association between the pre-/post-pandemic status and each record, possibly contributing to revealing the potential impacts of the pandemic on athletes’ preparedness for the Olympic games. Specifically, during our analysis, we will create graphs by using records in 2016 and 2020 Olympic Games and compare several models for the association between the pre-/post-pandemic status versus changes in marathon records.


### DATA
Our data was collected by web-scraping for 4 Olympic marathon games: Women/Men at Tokyo 2020 and Women/Men at Rio 2016. Athletes' names and results were obtained from result summaries, and further athletes' and their countries' backgrounds were obtained from additional resources. 

The publicly-available data sources we collected were from Wikipedia, COVID-19 Data Repository of Johns Hopkins University, the New World Encyclopedia (NWE), and World Bank.

The detailed codes are available below (1).

### EXPLORATORY ANALYSIS

  We used R to create nicely visualized graphs and build/compare different models. For visualization, we created **bar plots**, **box plots**, **histograms**, **scatter plots**, **stacked bar charts** and **maps**. We also created **animated line plots** using the Shiny app to see the animated differences of marathon finishing times between Rio 2016 and Tokyo 2020 across each rank. We also built **linear regression models** to examine the effect size of COVID-19 impact by using athlete finishing times as the outcome in addition to other covariates. In order to predict the outcome of Olympic marathon athlete placement in the Tokyo 2020 games, we performed **cutpoint analysis**, **machine learning** and **bootstrapping** to examine the accuracy of outcome prediction. 
  
The detailed codes are available below (2-6).


### FINAL ANALYSIS AND CONCLUSION
#### Data Scrapping
- Rio 2016: 155 male athletes from 79 countries (139 finished), 157 female athletes from 80 countries (133 finished). 
  
- Tokyo 2020: 106 male athletes from 46 countries (76 finished), 88 female athletes from 44 countries (73 finished) 

#### Visualization
- Despite a lower athlete attendance in the 2020 games, graph 3 and 4 showed that **the finishing times** in both men and women **all improved in Tokyo 2020** compared to Rio 2016. 

- However, there is **no strong relationship** observed between **COVID-19 severity** and **athletes' marathon finishing time** according to graph 9. Yet, graph 8 showed that **GDP**s of the countries where athletes originated from were **higher in 2020** among both the overall attending athletes and those who qualified the upper half, suggesting GDP could be a crucial factor in determining athlete's performance during the pandemic. 

- **We believe it is possible that higher-income countries were more capable of supporting marathon athletes to attend international competitions during the pandemic regardless of domestic COVID-19 severity**. 

#### Shiny
- **Tokyo 2020 had more faster records** than the Rio 2016 Olympics in each rank for both gender, and for the top few male records, Rio 2016 Olympics seemed to have a better performance; for the top few female records, Tokyo 2016 Olympics seemed to have a way better performance.

#### Linear Regression
- After **adjusting for** season best time, age, and continent, there is **no statistically significant difference** in marathon records between Rio and Tokyo Olympics for both men and women (α = 0.05).

- Also the multivariable linear regression showed that the association between marathon records and pre-/post COVID-19 Olympics **did not significantly differ** by the lockdown policy (α = 0.05).  

- However, the sample size of countries without the lockdown policy is small. Therefore, it is possible that we did not have enough power to detect a statistical interaction.

#### Machine Learning
- We used the optimal cutpoint **8239 (seconds)** for men, and the optimal cutpoint **9339 (seconds)** for women in our subsequent machine learning analyses.

- **Naive Bayes model** was the best model to predict male athletes' performance in the future Olympic Games during a similar pandemic. 

- **Knn model** was the best model to predict female athletes’ performance in the future Olympic Games during a similar pandemic. 

#### Bootrapping 
- We performed two sets of original vs. bootstrapped models: Decision tree vs. random forest and Naive Bayes vs. bootstrapped version of Naive Bayes. 

- Bootstrapping was most helpful in **the women case of tree and random forest**.
