==> 3_ShinyApp_YZ.Rmd <==
---
title: "Shiny App Section"
author: "Yuki Zhong (BST260 Group 11)"
output: html_document
runtime: shiny
---


### Shiny App Analysis


```{r}
# Libraries
library(tidyverse)
library(shiny)
library(ggthemes)
library(shinythemes)

# Read in dataframe
dat <- read_csv("data.csv")

shinyApp(ui = fluidPage( 
    # Change theme 
    theme = shinytheme("superhero"),
   
    #name title Panel
    titlePanel("Difference of time between 2016 and 2020 Olympics marathon record for each rank"),
    
    br(),
    
    # Sidebar
    sidebarLayout(
        
        # Widgets for selection
        sidebarPanel(
            # Explanatory text
            p("This Shiny apps allows you to use a slide bar to see the change of differene in each ranking between 2016 and 2020 Olympics, seperately for men and women."),
            
            br(),
            
            # Radio buttons that allows the user to choose a gender
            radioButtons(inputId = "sex", label = "Select men or women",
                         choices = c("Men", "Women")),
            
            # Input: rank slider with basic animation
            sliderInput("rank", "rank",
                        min = 1, max = 73,
                        value = 1, 
                        step = 1,
                        ticks = TRUE,  
                        animate = animationOptions(interval = 800) # add play speed
        )),
        
        # Main panel
        mainPanel(
          
            # Plot
            plotOutput("plot"),
            br(), 
            # Message about the difference in seconds for each rank
            textOutput("diff")
        )
    )
),

# Define server logic
server = function(input, output){
    
    # Make the selected line for the selected gender
    output$plot = renderPlot({
      
        # create the dataset for the plot
        i <- 1
        n <- input$rank+1
        y_diff<- c() #define an empty list that will be the y-value for the plot
        x_rank<- c() #define an empty list that will be the x-value for the plot
      
        #use a loop function to keep adding the previous differences and ranks to the list given a specific rank 
        while (i< n) {
          dat_diff <- dat %>% filter(sex == input$sex & rank == i) 
          y_diff <- c(y_diff,  diff(dat_diff$time_sec))
          x_rank <- c(x_rank, i)
          i <- i+1
        }

        # Scatterplot for the difference in seconds vs. rank
        ggplot() +
        theme_economist() +
        geom_line(aes(x = x_rank, y=y_diff), color = ifelse(input$sex == "Men", 'deepskyblue2','hotpink2'), size = 1) +
        xlab("rank") +
        ylab("Difference (seconds Rio - Tokyo)") +
        geom_hline(yintercept = 0, colour = 'black') 
        
            
    })
    
    # Identify which Olympics had a faster record by how many seconds, for each rank and gender
    output$diff = renderText({
        dat_diff <- dat %>% filter(sex == input$sex & rank == input$rank) 
        paste0("For rank ", input$rank, " in ",input$sex, "'s Marathon , ", ifelse(diff(dat_diff$time_sec)<0,  "the record in Tokyo 2020", "the record in Rio 2016"), " is ", abs(diff(dat_diff$time_sec)), " seconds faster")
    })
})
```


**Interpretation:**
From the interactive graphs, we can see that for both men and women, Tokyo 2020 Olympics seemed to have faster records overall. For men, the records in Rio are faster in high ranks between 1-16 and middle ranks between 30-40. For women, the records in Rio are faster only sparsely between the rank 30 and 50. In addition, we can see that the difference of time is more exaggerated in high ranks and low ranks, while the differences in the middle ranks are not too many.
